+++ src/main/java/com/dejvo/Shop/db/mapper/SellerRowMapper.java
@@ -11,7 +11,7 @@ public class SellerRowMapper implements RowMapper<Seller> {
     @Override
     public Seller mapRow(ResultSet resultSet, int i) throws SQLException {
         Seller seller= new Seller();
-        seller.setId(resultSet.getLong("id"));
+        seller.setId(resultSet.getInt("id"));
         seller.setName(resultSet.getString("name"));
         seller.setEmail(resultSet.getString("email"));
         seller.setAddress(resultSet.getString("address"));
